Change the way card strings are read
Now all the strings will be read as( (int32)code<<32)|((int32) offset), where if the code is 0, that'll be a system string

+++ gframe/data_manager.h
@@ -26,7 +26,7 @@ public:
 	std::wstring GetName(int code);
 	std::wstring GetText(int code);
 	std::wstring GetDesc(u64 strCode);
-	std::wstring GetSysString(int code);
+	std::wstring GetSysString(u64 code);
 	std::wstring GetVictoryString(int code);
 	std::wstring GetCounterName(int code);
 	std::wstring GetSetName(int code);
